Flyttet input-kontroll av datoer ut i en felles-modul (util). Ny versjon av Kotlin
