+++ pom.xml
@@ -26,7 +26,7 @@
 
     <!-- language -->
     <java.version>14</java.version>
-    <kotlin.version>1.3.72</kotlin.version>
+    <kotlin.version>1.4.0</kotlin.version>
 
     <!-- build -->
     <build-helper-maven.version>3.1.0</build-helper-maven.version>
@@ -96,7 +96,7 @@
 
     <dependency>
       <groupId>org.jetbrains.kotlin</groupId>
-      <artifactId>kotlin-test</artifactId>
+      <artifactId>kotlin-test-junit</artifactId>
       <version>${kotlin.version}</version>
       <scope>test</scope>
     </dependency>
